chore(RHTAPWATCH-691): verify-rpms test to fail on unsigned RPMs
The e2e test for the verify-rpms script checked
only a scenario where it should not fail when unsigned RPMs
or errors exists.

Adds tests for multiple scenarios - where the script should
and should not fail when encountering unsigned RPMs or other errors.

This lead to a refactor which changed the failing responsability
from which fails the process from "generate_output" to the
main function.

Signed-off-by: Omer <oamsalem@redhat.com>